Add additional code part after step 1 to have code injection in the documentation
