+++ tutorials/tour-of-restate-typescript/README.md
@@ -29,6 +29,7 @@ npm run app
 npm run part1
 npm run part2
 npm run part3
-npm run part4
+npm run part4 
+npm run part5
 ```
