--- tutorials/tour-of-restate-typescript/src/part1b/checkout.ts
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2024 - Restate Software, Inc., Restate GmbH
- *
- * This file is part of the Restate examples,
- * which is released under the MIT license.
- *
- * You can find a copy of the license in the file LICENSE
- * in the root directory of this repository or package or at
- * https://github.com/restatedev/examples/
- */
-
-import * as restate from "@restatedev/restate-sdk";
-
-// <start_checkout>
-export const checkoutRouter = restate.router({
-  async handle(ctx: restate.Context, request: { userId: string; tickets: string[] }){
-    return true;
-  },
-});
-// <end_checkout>
-
-export const checkoutApi: restate.ServiceApi<typeof checkoutRouter> = {
-  path: "Checkout",
-};
