+++ tutorials/tour-of-restate-typescript/src/part2/user_session.ts
@@ -11,7 +11,9 @@
 
 import * as restate from "@restatedev/restate-sdk";
 import { ticketServiceApi } from "./ticket_service";
+// <start_checkout_api_import>
 import { checkoutApi } from "./checkout";
+// <end_checkout_api_import>
 
 export const userSessionRouter = restate.keyedRouter({
   // <start_add_ticket>
@@ -19,12 +21,7 @@ export const userSessionRouter = restate.keyedRouter({
     const reservationSuccess = await ctx.rpc(ticketServiceApi).reserve(ticketId);
 
     if (reservationSuccess) {
-      //highlight-next-line
-      const tickets = (await ctx.get<string[]>("tickets")) ?? [];
-      tickets.push(ticketId);
-      //highlight-next-line
-      ctx.set("tickets", tickets);
-
+      // highlight-next-line
       ctx.sendDelayed(userSessionApi, 15 * 60 * 1000).expireTicket(userId, ticketId);
     }
 
@@ -32,42 +29,24 @@ export const userSessionRouter = restate.keyedRouter({
   },
   // <end_add_ticket>
 
-  // <start_expire_ticket>
-  async expireTicket(ctx: restate.KeyedContext, userId: string, ticketId: string){
-    const tickets = (await ctx.get<string[]>("tickets")) ?? [];
-
-    const ticketIndex = tickets.findIndex((ticket) => ticket === ticketId);
-
-    if (ticketIndex != -1) {
-      tickets.splice(ticketIndex, 1);
-      ctx.set("tickets", tickets);
-
-      ctx.send(ticketServiceApi).unreserve(ticketId);
-    }
-  },
-  // <end_expire_ticket>
-
   // <start_checkout>
-  async checkout(ctx: restate.KeyedContext, userId: string){
-    //highlight-next-line
-    const tickets = (await ctx.get<string[]>("tickets")) ?? [];
-
+  async checkout(ctx: restate.KeyedContext, userId: string) {
     //highlight-start
-    if (tickets.length === 0) {
-      return false;
-    }
+    const checkoutRequest = { userId: userId, tickets: ["456"] };
+    const success = await ctx.rpc(checkoutApi).handle(checkoutRequest);
     //highlight-end
 
-    const checkoutSuccess = await ctx.rpc(checkoutApi)
-        .handle({ userId: userId, tickets: tickets });
-
-    if (checkoutSuccess) {
-      ctx.clear("tickets");
-    }
-
-    return checkoutSuccess;
+    return success;
   },
   // <end_checkout>
+
+  // <start_expire_ticket>
+  async expireTicket(ctx: restate.KeyedContext, userId: string, ticketId: string){
+    // highlight-start
+    ctx.send(ticketServiceApi).unreserve(ticketId);
+    // highlight-end
+  }
+  // <end_expire_ticket>
 });
 
 export const userSessionApi: restate.ServiceApi<typeof userSessionRouter> = {
