+++ tutorials/tour-of-restate-typescript/src/part3/checkout.ts
@@ -10,35 +10,12 @@
  */
 
 import * as restate from "@restatedev/restate-sdk";
-import { PaymentClient } from "../auxiliary/payment_client";
-import { EmailClient } from "../auxiliary/email_client";
 
-// <start_checkout>
 export const checkoutRouter = restate.router({
   async handle(ctx: restate.Context, request: { userId: string; tickets: string[] }){
-    // <start_side_effects>
-    // We are a uniform shop where everything costs 40 USD
-    const totalPrice = request.tickets.length * 40;
-    const paymentClient = PaymentClient.get();
-
-    // highlight-start
-    const idempotencyKey = ctx.rand.uuidv4();
-    const success = await ctx.sideEffect(() => paymentClient.call(idempotencyKey, totalPrice));
-    // highlight-end
-    // <end_side_effects>
-
-    const email = EmailClient.get();
-
-    if (success) {
-      await ctx.sideEffect(() => email.notifyUserOfPaymentSuccess(request.userId));
-    } else {
-      await ctx.sideEffect(() => email.notifyUserOfPaymentFailure(request.userId));
-    }
-
-    return success;
+    return true;
   },
 });
-// <end_checkout>
 
 export const checkoutApi: restate.ServiceApi<typeof checkoutRouter> = {
   path: "Checkout",
