+++ tutorials/tour-of-restate-typescript/src/part3/ticket_service.ts
@@ -18,6 +18,7 @@ enum TicketStatus {
 }
 
 export const ticketDbRouter = restate.keyedRouter({
+  // <start_reserve>
   async reserve(ctx: restate.KeyedContext){
     const status =
       (await ctx.get<TicketStatus>("status")) ?? TicketStatus.Available;
@@ -29,8 +30,10 @@ export const ticketDbRouter = restate.keyedRouter({
       return false;
     }
   },
+  // <end_reserve>
 
-  async unreserve(ctx: restate.KeyedContext){
+  // <start_unreserve>
+  async unreserve(ctx: restate.KeyedContext) {
     const status =
       (await ctx.get<TicketStatus>("status")) ?? TicketStatus.Available;
 
@@ -38,7 +41,9 @@ export const ticketDbRouter = restate.keyedRouter({
       ctx.clear("status");
     }
   },
+  // <end_unreserve>
 
+  // <start_mark_as_sold>
   async markAsSold(ctx: restate.KeyedContext){
     const status =
       (await ctx.get<TicketStatus>("status")) ?? TicketStatus.Available;
@@ -47,6 +52,7 @@ export const ticketDbRouter = restate.keyedRouter({
       ctx.set("status", TicketStatus.Sold);
     }
   },
+  // <end_mark_as_sold>
 });
 
 export const ticketServiceApi: restate.ServiceApi<typeof ticketDbRouter> = {
