+++ tutorials/tour-of-restate-typescript/src/part5/checkout.ts
@@ -10,35 +10,42 @@
  */
 
 import * as restate from "@restatedev/restate-sdk";
-import { PaymentClient } from "../auxiliary/payment_client";
+import { v4 as uuid } from "uuid";
 import { EmailClient } from "../auxiliary/email_client";
+import { PaymentClient } from "../auxiliary/payment_client";
 
-// <start_checkout>
 export const checkoutRouter = restate.router({
   async handle(ctx: restate.Context, request: { userId: string; tickets: string[] }){
-    // <start_side_effects>
+    // Generate idempotency key for the stripe client
+    const idempotencyKey = await ctx.sideEffect(async () => uuid());
+
     // We are a uniform shop where everything costs 40 USD
     const totalPrice = request.tickets.length * 40;
-    const paymentClient = PaymentClient.get();
 
-    // highlight-start
-    const idempotencyKey = ctx.rand.uuidv4();
-    const success = await ctx.sideEffect(() => paymentClient.call(idempotencyKey, totalPrice));
-    // highlight-end
-    // <end_side_effects>
+    // <start_failing_client>
+    const paymentClient = PaymentClient.get();
+    //highlight-start
+    const doPayment = () =>
+      paymentClient.failingCall(idempotencyKey, totalPrice);
+    const success = await ctx.sideEffect(doPayment);
+    //highlight-end
+    // <end_failing_client>
 
     const email = EmailClient.get();
 
     if (success) {
-      await ctx.sideEffect(() => email.notifyUserOfPaymentSuccess(request.userId));
+      await ctx.sideEffect(() =>
+        email.notifyUserOfPaymentSuccess(request.userId),
+      );
     } else {
-      await ctx.sideEffect(() => email.notifyUserOfPaymentFailure(request.userId));
+      await ctx.sideEffect(() =>
+        email.notifyUserOfPaymentFailure(request.userId),
+      );
     }
 
     return success;
   },
 });
-// <end_checkout>
 
 export const checkoutApi: restate.ServiceApi<typeof checkoutRouter> = {
   path: "Checkout",
