+++ tutorials/tour-of-restate-typescript/src/part5/ticket_service.ts
@@ -10,9 +10,6 @@
  */
 
 import * as restate from "@restatedev/restate-sdk";
-// <start_import_sleep>
-import { setTimeout } from "timers/promises";
-// <end_import_sleep>
 
 enum TicketStatus {
   Available,
@@ -21,21 +18,34 @@ enum TicketStatus {
 }
 
 export const ticketDbRouter = restate.keyedRouter({
-  // <start_reserve>
   async reserve(ctx: restate.KeyedContext){
-    //good-code-start
-    await ctx.sleep(35000);
-    //good-code-end
-    return true;
+    const status =
+      (await ctx.get<TicketStatus>("status")) ?? TicketStatus.Available;
+
+    if (status === TicketStatus.Available) {
+      ctx.set("status", TicketStatus.Reserved);
+      return true;
+    } else {
+      return false;
+    }
   },
-  // <end_reserve>
 
   async unreserve(ctx: restate.KeyedContext){
-    return true;
+    const status =
+      (await ctx.get<TicketStatus>("status")) ?? TicketStatus.Available;
+
+    if (status !== TicketStatus.Sold) {
+      ctx.clear("status");
+    }
   },
 
   async markAsSold(ctx: restate.KeyedContext){
-    return true;
+    const status =
+      (await ctx.get<TicketStatus>("status")) ?? TicketStatus.Available;
+
+    if (status === TicketStatus.Reserved) {
+      ctx.set("status", TicketStatus.Sold);
+    }
   },
 });
