+++ tutorials/tour-of-restate-typescript/src/part5/user_session.ts
@@ -14,60 +14,60 @@ import { ticketServiceApi } from "./ticket_service";
 import { checkoutApi } from "./checkout";
 
 export const userSessionRouter = restate.keyedRouter({
-  // <start_add_ticket>
   async addTicket(ctx: restate.KeyedContext, userId: string, ticketId: string){
-    const reservationSuccess = await ctx.rpc(ticketServiceApi).reserve(ticketId);
+    // try to reserve ticket
+    const reservation_success = await ctx.rpc(ticketServiceApi).reserve(ticketId);
 
-    if (reservationSuccess) {
-      //highlight-next-line
+    if (reservation_success) {
+      // add ticket to user session tickets
       const tickets = (await ctx.get<string[]>("tickets")) ?? [];
       tickets.push(ticketId);
-      //highlight-next-line
       ctx.set("tickets", tickets);
 
-      ctx.sendDelayed(userSessionApi, 15 * 60 * 1000).expireTicket(userId, ticketId);
+      // Schedule expiry timer
+      ctx
+        .sendDelayed(userSessionApi, 15 * 60 * 1000)
+        .expireTicket(userId, ticketId);
     }
 
-    return reservationSuccess;
+    return reservation_success;
   },
-  // <end_add_ticket>
 
-  // <start_expire_ticket>
   async expireTicket(ctx: restate.KeyedContext, userId: string, ticketId: string){
     const tickets = (await ctx.get<string[]>("tickets")) ?? [];
 
-    const ticketIndex = tickets.findIndex((ticket) => ticket === ticketId);
+    const index = tickets.findIndex((id) => id === ticketId);
 
-    if (ticketIndex != -1) {
-      tickets.splice(ticketIndex, 1);
+    // try removing ticket
+    if (index != -1) {
+      tickets.splice(index, 1);
       ctx.set("tickets", tickets);
-
+      // unreserve if ticket was reserved before
       ctx.send(ticketServiceApi).unreserve(ticketId);
     }
   },
-  // <end_expire_ticket>
 
-  // <start_checkout>
   async checkout(ctx: restate.KeyedContext, userId: string){
-    //highlight-next-line
     const tickets = (await ctx.get<string[]>("tickets")) ?? [];
 
-    //highlight-start
     if (tickets.length === 0) {
       return false;
     }
-    //highlight-end
 
-    const checkoutSuccess = await ctx.rpc(checkoutApi)
-        .handle({ userId: userId, tickets: tickets });
+    const checkoutSuccess = await ctx
+      .rpc(checkoutApi)
+      .handle({ userId: userId, tickets: tickets! });
 
     if (checkoutSuccess) {
+      // mark tickets as sold if checkout was successful
+      for (const ticketId of tickets) {
+        ctx.send(ticketServiceApi).markAsSold(ticketId);
+      }
       ctx.clear("tickets");
     }
 
     return checkoutSuccess;
   },
-  // <end_checkout>
 });
 
 export const userSessionApi: restate.ServiceApi<typeof userSessionRouter> = {
